refactor(app,desktop)!: 删掉本地 leaderboard 统计（.teamclu/stats.json）
团队 leaderboard 早就全云端了——LeaderboardSection → fetchTeamLeaderboard →
telemetry.listLeaderboard(teamId, period)，读的是云端聚合。本地那套只剩写：

- `loadStats` 唯一的调用者是打开 workspace 时的 loadLocalStatsForWorkspace，加载完
  存进 store，而没有任何组件渲染 store 里的 stats —— 实际是只写不读。
- 4 处写入（telemetry 的 feedback / star / session 完成，MqttLiveWiring 的 skill
  调用）每一处都紧跟一次云端写入，所以删掉不丢数据。
- `exportTeamData` / `scheduleTeamFeedbackExport` / `triggerTeamLeaderboardExport`
  三个都已经是空函数（注释：Leaderboard now lives in the public.team_leaderboard
  view），但 workspace.ts 仍每次打开工作区调 exportTeamData，注释还写着"ensure
  .leaderboard directory exists"——一并删掉。

唯一不能跟着删的是 skill 上报：incrementSkillUsage 除了写本地，还会
insertSkillUsage 到云端，那是 leaderboard 技能维度的唯一来源。抽成
lib/telemetry/skill-usage.ts 的 reportSkillUsage，由 MqttLiveWiring 直接调用，
原来针对云端半边的 4 个用例搬到 __tests__/skill-usage.test.ts。

删除面：stores/local-stats.ts、lib/local-stats/types.ts、
commands/local_stats.rs（4 个 Tauri command + lib.rs 注册 + mod 声明）、
stores/__tests__/local-stats-skills.test.ts。

`.teamclu/stats.json` 从此不再被创建；已存在的文件留在磁盘上不管（不主动删用户
数据），反正没人再读。

验证：packages/app tsc 通过、vitest 全量 446 files / 2881 tests 全绿；
apps/desktop cargo check 通过。
