+++ apps/desktop/src/lib.rs
@@ -504,10 +504,6 @@ pub fn run() {
             commands::env_vars::env_catalog_delete,
             commands::env_vars::env_var_resolve,
             commands::session_export::session_export,
-            commands::local_stats::read_local_stats,
-            commands::local_stats::write_local_stats,
-            commands::local_stats::update_local_stats,
-            commands::local_stats::reset_local_stats,
             local_cache::commands::local_cache_actor_upsert_batch,
             local_cache::commands::local_cache_actor_load_team,
             local_cache::commands::local_cache_actor_load_by_ids,
