--- packages/app/src/lib/local-stats/types.ts
@@ -1,48 +1,0 @@
-/**
- * Local statistics types for project-level tracking
- * Stored in .teamclu/stats.json (not uploaded, project-specific)
- */
-
-export interface LocalStats {
-  version: string
-  taskCompleted: number      // Number of completed tasks/sessions
-  totalTokens: number         // Total tokens consumed
-  totalCost: number          // Total cost in USD
-  feedbackCount: number      // Total feedback count
-  positiveCount: number      // Positive feedback count
-  negativeCount: number      // Negative feedback count
-  starRatings: StarRatings   // Star rating distribution
-  sessions: SessionStats     // Session statistics
-  lastUpdated: string        // ISO 8601 timestamp
-  createdAt: string          // ISO 8601 timestamp
-  skillUsage: Record<string, number>  // Skill usage count by skill name
-}
-
-export interface StarRatings {
-  1: number
-  2: number
-  3: number
-  4: number
-  5: number
-}
-
-export interface SessionStats {
-  total: number              // Total number of sessions
-  withFeedback: number       // Sessions with at least one feedback
-}
-
-export interface LocalStatsUpdate {
-  taskCompleted?: number
-  totalTokens?: number
-  totalCost?: number
-  feedbackCount?: number
-  positiveCount?: number
-  negativeCount?: number
-  starRating?: 1 | 2 | 3 | 4 | 5
-  sessionsTotal?: number
-  sessionsWithFeedback?: number
-  skillInvoked?: string
-}
-
-export type FeedbackRating = 'positive' | 'negative'
-export type StarRating = 1 | 2 | 3 | 4 | 5
